+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Browse view now shows a message when a root is being scanned instead of displaying "No items found"
+
 ## [v0.2.1] - 2025-11-08
 
 ### Breaking Changes